# doc/openapi/swagger.yaml
basePath: /
definitions:
...
  model.Data:
    properties:
      message:
        type: string
    type: object
  model.HelloJsonResponse:
    properties:
      data:
        $ref: '#/definitions/model.Data'
    type: object
...
info:
  contact: {}
  description: This is a sample API server using Gin and Swagger.
  title: My API
  version: "1.0"
paths:
...
  /hello/json:
    get:
      consumes:
      - application/json
      description: get hello world json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HelloJsonResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get hello world json
      tags:
      - hello
  /hello/string:
    get:
      consumes:
      - application/json
      description: get hello world string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get hello world string
      tags:
      - hello
...
schemes:
- http
- https
swagger: "2.0"